change class back to default export from package
